pyproject.toml
[tool.poetry]
name = "wiremock-lambda-mock"
version = "0.1.0"
description = "WireMock-based local mocking for AWS Lambda invoke endpoints with JSON stub templates"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
wiremock = ">=2.6.1,<3.0.0"
requests = ">=2.31.0"
PyYAML = ">=6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4"

[tool.pytest.ini_options]
addopts = "-q"
python_files = ["tests/test_*.py"]

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
